Moved namespace gpu_reconstruction_kernels defination in GPUReconstruction class
