+++ GPU/GPUTracking/Base/GPUReconstruction.h
@@ -63,8 +63,37 @@ struct GPUMemoryReuse;
 
 namespace gpu_reconstruction_kernels
 {
-struct deviceEvent;
-class threadContext;
+struct deviceEvent {
+  constexpr deviceEvent() = default;
+  constexpr deviceEvent(std::nullptr_t p) : v(nullptr){};
+  template <class T>
+  void set(T val)
+  {
+    v = reinterpret_cast<void*&>(val);
+  }
+  template <class T>
+  T& get()
+  {
+    return reinterpret_cast<T&>(v);
+  }
+  template <class T>
+  T* getEventList()
+  {
+    return reinterpret_cast<T*>(this);
+  }
+  bool isSet() const { return v; }
+
+ private:
+  void* v = nullptr; // We use only pointers anyway, and since cl_event and cudaEvent_t and hipEvent_t are actually pointers, we can cast them to deviceEvent (void*) this way.
+};
+
+class threadContext
+{
+ public:
+  threadContext();
+  virtual ~threadContext();
+};
+
 } // namespace gpu_reconstruction_kernels
 
 class GPUReconstruction
